Implement composite keys for boundaries
This commit introduces support for composite keys in boundary queries,
allowing multiple arguments to be used to lookup an object. For example:

```
type Query {
  product(shopId: ID!, handle: String!): Product @stitch(key: "shopId handle")
}
```

This will select both the `shopId` and `handle` fields from the
`Product` type in one schema to be used to lookup the same object in
another schema.